+++ src/main/java/com/chaos131/ctre/ChaosCanCoder.java
@@ -2,8 +2,10 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package com.chaos131.util;
+package com.chaos131.ctre;
 
+import com.chaos131.can.CanConstants.CanBusName;
+import com.chaos131.can.CanConstants.CanId;
 import com.ctre.phoenix6.configs.CANcoderConfiguration;
 import com.ctre.phoenix6.hardware.CANcoder;
 
@@ -12,8 +14,8 @@ public class ChaosCanCoder extends CANcoder {
   public final CANcoderConfiguration Configuration = new CANcoderConfiguration();
 
   /** Creates the new TalonFX wrapper WITHOUT simulation support. */
-  public ChaosCanCoder(int canId, String canBus) {
-    super(canId, canBus);
+  public ChaosCanCoder(CanId canId, CanBusName canBus) {
+    super(canId.id, canBus.name);
   }
 
   /** Applies/burns the configuration to the motor. */
