+++ src/main/java/com/chaos131/ctre/ChaosTalonFxs.java
@@ -2,11 +2,13 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package com.chaos131.util;
+package com.chaos131.ctre;
 
 import static edu.wpi.first.units.Units.Radians;
 import static edu.wpi.first.units.Units.Volts;
 
+import com.chaos131.can.CanConstants.CanBusName;
+import com.chaos131.can.CanConstants.CanId;
 import com.chaos131.pid.PIDFValue;
 import com.ctre.phoenix6.configs.Slot0Configs;
 import com.ctre.phoenix6.configs.TalonFXSConfiguration;
@@ -34,8 +36,8 @@ public class ChaosTalonFxs extends TalonFXS {
   public final TalonFXSConfiguration Configuration = new TalonFXSConfiguration();
 
   /** Creates the new TalonFX wrapper WITHOUT simulation support. */
-  public ChaosTalonFxs(int canId, String canBus) {
-    super(canId, canBus);
+  public ChaosTalonFxs(CanId canId, CanBusName canBus) {
+    super(canId.id, canBus.name);
     this.m_gearRatio = 0.0;
     m_motorSimModel = null;
     m_isMainSimMotor = false;
